Move signup notification signal to models and tests
