+++ django_project/gap/models/signup_request.py
@@ -10,6 +10,10 @@ from django.contrib.gis.db import models
 from django.conf import settings
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
+from django.db.models.signals import post_save
+from django.dispatch import receiver
+from django.core.mail import send_mail
+from django.contrib.auth.models import Group
 
 
 class RequestStatus(models.TextChoices):
@@ -71,3 +75,31 @@ class SignUpRequest(models.Model):
 
     def __str__(self):
         return f'{self.first_name} {self.last_name} <{self.email}>'
+
+
+@receiver(post_save, sender=SignUpRequest)
+def notify_user_managers_on_signup(sender, instance, created, **kwargs):
+    """Notify user managers on sign up."""
+    if not created:
+        return
+
+    try:
+        group = Group.objects.get(name="User Manager")
+        managers = group.user_set.filter(is_active=True, email__isnull=False)
+
+        recipient_list = [user.email for user in managers if user.email]
+        if recipient_list:
+            send_mail(
+                subject="New Sign Up Request",
+                message=(
+                    f"A new sign-up request has been submitted:\n\n"
+                    f"Name: {instance.first_name} {instance.last_name}\n"
+                    f"Email: {instance.email}\n"
+                    f'Description: {instance.description or "-"}'
+                ),
+                from_email=settings.DEFAULT_FROM_EMAIL,
+                recipient_list=recipient_list,
+                fail_silently=False,
+            )
+    except Group.DoesNotExist:
+        pass
